Shape and color buttons are working!
Updated the files so that when a color or shape icon is clicked, an alert shows up saying that each (respective) button has been pressed. These buttons have been added to the menu bar on the left side. Next big step will be adding an editor component on the screen when one of the buttons is clicked, and being able to change the color of them.