Managed to add element to HTML with JQuery!
Next goal is to make this conditional and put it in organize.js. If "class" is detected in the JSON file, then make one of these elements in the UI. Figure out why the menu components keep shifting sometimes. Also practiced with JS classes, which I'll probably end up using.